+++ NEWS.md
@@ -1,4 +1,5 @@
 # lintr 1.0.3.9000 # 
+* Added the new `paren_brace_linter()` to the default linters; it checks that there is a space between right parenthesis and an opening curly brace (@bfgray3, #242).
 * Now lintr only tries to generate comments if running in wercker or travis CI (#166)
 * Top level calls that contain function definitions are not properly ignored by
   the object usage linter (#26).